Added a routes file. updated ensure authentication function
