+++ server/userProfile/userOAuthUtils.js
@@ -4,7 +4,7 @@ var GitHubStrategy = require('passport-github').Strategy;
 var session = require('express-session');
 var db = require('./UserModel');
 var User = db.user;
-
+var globalProfile; 
 // Passport session setup.
 //   To support persistent login sessions, Passport needs to be able to
 //   serialize users into and deserialize users out of the session.  Typically,
